+++ lib/widgets/quickalert_dialog.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:quickalert/models/quickalert_animtype.dart';
 import 'package:quickalert/models/quickalert_options.dart';
 import 'package:quickalert/models/quickalert_type.dart';
@@ -128,9 +129,20 @@ class QuickAlert {
       customAsset: customAsset,
       width: width,
     );
-    
-    final child = WillPopScope(
-      onWillPop: () => Future.value(!disableBackBtn),
+
+    Widget child = WillPopScope(
+      onWillPop: () async {
+        options.timer?.cancel();
+        if (options.type == QuickAlertType.loading &&
+            !disableBackBtn &&
+            showCancelBtn) {
+          if (options.onCancelBtnTap != null) {
+            options.onCancelBtnTap!();
+            return false;
+          }
+        }
+        return !disableBackBtn;
+      },
       child: AlertDialog(
         contentPadding: EdgeInsets.zero,
         shape: RoundedRectangleBorder(
@@ -141,6 +153,23 @@ class QuickAlert {
         ),
       ),
     );
+
+    if (options.type != QuickAlertType.loading) {
+      child = RawKeyboardListener(
+        focusNode: FocusNode(),
+        autofocus: true,
+        onKey: (event) {
+          if (event is RawKeyUpEvent &&
+              event.logicalKey == LogicalKeyboardKey.enter) {
+            options.timer?.cancel();
+            options.onConfirmBtnTap != null
+                ? options.onConfirmBtnTap!()
+                : Navigator.pop(context);
+          }
+        },
+        child: child,
+      );
+    }
 
     return showGeneralDialog(
       barrierColor: barrierColor ?? Colors.black.withOpacity(0.5),